+++ shuup/front/apps/registration/views.py
@@ -53,10 +53,7 @@ class RegistrationViewMixin(object):
 
     def register(self, form):
         user = super(RegistrationViewMixin, self).register(form)
-
-        if settings.SHUUP_ENABLE_MULTIPLE_SHOPS and settings.SHUUP_MANAGE_CONTACTS_PER_SHOP:
-            get_person_contact(user).shops.add(self.request.shop)
-
+        get_person_contact(user).add_to_shop(self.request.shop)
         return user
 
 
@@ -94,7 +91,7 @@ class CompanyRegistrationView(RegistrationViewMixin, default_views.RegistrationV
 
         if settings.SHUUP_ENABLE_MULTIPLE_SHOPS and settings.SHUUP_MANAGE_CONTACTS_PER_SHOP:
             company = get_company_contact(user)
-            company.shops.add(self.request.shop)
+            company.add_to_shop(self.request.shop)
 
 
 class ActivationView(default_views.ActivationView):