+++ shuup_tests/tasks/test_admin_module.py
@@ -33,7 +33,7 @@ def test_task_admin(admin_user):
     admin_user.set_password(ADMIN_PWD)
     admin_user.save()
     admin_contact = get_person_contact(admin_user)
-    admin_contact.shops.add(shop)
+    admin_contact.add_to_shop(shop)
 
     client = SmartClient()
     client.login(username=admin_user.username, password=ADMIN_PWD)
@@ -152,7 +152,7 @@ def test_task_type_admin(admin_user):
     admin_user.set_password(ADMIN_PWD)
     admin_user.save()
     admin_contact = get_person_contact(admin_user)
-    admin_contact.shops.add(shop)
+    admin_contact.add_to_shop(shop)
 
     client = SmartClient()
     client.login(username=admin_user.username, password=ADMIN_PWD)